Loopline/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
black==25.9.0
certifi==2025.1.31
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-anymail==14.0
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
Faker==37.8.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-django==4.11.1
python-dotenv==1.1.0
pytokens==0.1.10
redis==6.3.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
zope.interface==7.2
